Preserve priorities for OF1.0 exact-match flows.

+++ lib/ofp-actions.h
@@ -548,7 +548,7 @@ struct ofpact_learn {
 
     uint16_t idle_timeout;      /* Idle time before discarding (seconds). */
     uint16_t hard_timeout;      /* Max time before discarding (seconds). */
-    uint16_t priority;          /* Priority level of flow entry. */
+    int priority;               /* Priority level of flow entry. */
     uint8_t table_id;           /* Table to insert flow entry. */
     ovs_be64 cookie;            /* Cookie for new flow. */
     enum nx_learn_flags flags;  /* NX_LEARN_F_*. */
